requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
array-record==0.2.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorful==0.5.5
contourpy==1.0.7
cycler==0.11.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distlib==0.3.6
dm-tree==0.1.8
etils==1.3.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
gast==0.4.0
google-api-core==2.11.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
gpustat==1.1
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.29.0
importlib-resources==5.12.0
ipython @ file:///croot/ipython_1680701871216/work
itsdangerous==2.1.2
jax==0.4.10
jedi @ file:///work/ci_py311_2/jedi_1679336495545/work
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kaggle==1.5.15
keras==2.12.0
keras-vis==0.4.1
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.1.23
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
platformdirs==3.5.1
prettytable==3.9.0
prometheus-client==0.17.0
promise==2.3
prompt-toolkit @ file:///work/ci_py311/prompt-toolkit_1676824857154/work
protobuf==4.23.1
psutil==5.9.5
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
python-version==0.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
ray==2.4.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorrt==8.6.1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets @ file:///work/ci_py311/traitlets_1676823305040/work
triton==2.0.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.21.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.3.4
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0